+++ src/parquet/schema.cc
@@ -638,7 +638,8 @@ void SchemaDescriptor::BuildTree(const NodePtr& node, int16_t max_def_level,
     // Primitive node, append to leaves
     leaves_.push_back(ColumnDescriptor(node, max_def_level, max_rep_level, this));
     leaf_to_base_.emplace(static_cast<int>(leaves_.size()) - 1, base);
-    leaf_to_idx_.emplace(node->path()->ToDotString(), leaves_.size() - 1);
+    leaf_to_idx_.emplace(
+        node->path()->ToDotString(), static_cast<int>(leaves_.size()) - 1);
   }
 }
